Add endpoint to fetch user's watch history
The code changes in the `user.controller.js` file add a new endpoint `getWatchHistory` to fetch the watch history of a user. This endpoint uses aggregation pipelines to perform a lookup on the `videos` collection and retrieve the user's watch history along with additional information about the videos and their owners.